all: Add RU_864_870 band definition
