+++ pkg/band/ru_864_870.go
@@ -1,4 +1,4 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,21 +14,15 @@
 
 package band
 
-import (
-	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+import "go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
+
+const (
+	// RU_864_870 is the ID of the Russian frequency plan
+	RU_864_870 = "RU_864_870"
 )
 
-//revive:disable:var-naming
-
-var ru_864_870 Band
-
-// RU_864_870 is the ID of the Russian frequency plan
-const RU_864_870 = "RU_864_870"
-
-//revive:enable:var-naming
-
-func init() {
-	defaultChannels := []Channel{
+var (
+	ru864870DefaultChannels = []Channel{
 		{
 			Frequency:   868900000,
 			MaxDataRate: ttnpb.DATA_RATE_5,
@@ -39,7 +33,7 @@ func init() {
 		},
 	}
 
-	downlinkDRTable := [8][6]ttnpb.DataRateIndex{
+	ru864870DownlinkDRTable = [8][6]ttnpb.DataRateIndex{
 		{0, 0, 0, 0, 0, 0},
 		{1, 0, 0, 0, 0, 0},
 		{2, 1, 0, 0, 0, 0},
@@ -49,96 +43,4 @@ func init() {
 		{6, 5, 4, 3, 2, 1},
 		{7, 6, 5, 4, 3, 2},
 	}
-
-	ru_864_870 = Band{
-		ID: RU_864_870,
-
-		EnableADR: true,
-
-		MaxUplinkChannels: 16,
-		UplinkChannels:    defaultChannels,
-
-		MaxDownlinkChannels: 16,
-		DownlinkChannels:    defaultChannels,
-
-		// MaxTxPower as per Appendix 10 to the decision of GKRCh from 07.05. 2007
-		SubBands: []SubBandParameters{
-			{
-				MinFrequency: 864000000,
-				MaxFrequency: 870000000,
-				DutyCycle:    0.01,
-				MaxEIRP:      20.0 + eirpDelta,
-			},
-		},
-
-		DataRates: map[ttnpb.DataRateIndex]DataRate{
-			ttnpb.DATA_RATE_0: makeLoRaDataRate(12, 125000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_1: makeLoRaDataRate(11, 125000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_2: makeLoRaDataRate(10, 125000, makeConstMaxMACPayloadSizeFunc(59)),
-			ttnpb.DATA_RATE_3: makeLoRaDataRate(9, 125000, makeConstMaxMACPayloadSizeFunc(123)),
-			ttnpb.DATA_RATE_4: makeLoRaDataRate(8, 125000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_5: makeLoRaDataRate(7, 125000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_6: makeLoRaDataRate(7, 250000, makeConstMaxMACPayloadSizeFunc(230)),
-			ttnpb.DATA_RATE_7: makeFSKDataRate(50000, makeConstMaxMACPayloadSizeFunc(230)),
-		},
-		MaxADRDataRateIndex: ttnpb.DATA_RATE_5,
-
-		ReceiveDelay1:        defaultReceiveDelay1,
-		ReceiveDelay2:        defaultReceiveDelay2,
-		JoinAcceptDelay1:     defaultJoinAcceptDelay1,
-		JoinAcceptDelay2:     defaultJoinAcceptDelay2,
-		MaxFCntGap:           defaultMaxFCntGap,
-		ADRAckLimit:          defaultADRAckLimit,
-		ADRAckDelay:          defaultADRAckDelay,
-		MinRetransmitTimeout: defaultRetransmitTimeout - defaultRetransmitTimeoutMargin,
-		MaxRetransmitTimeout: defaultRetransmitTimeout + defaultRetransmitTimeoutMargin,
-
-		DefaultMaxEIRP: 16,
-		TxOffset: []float32{
-			0,
-			-2,
-			-4,
-			-6,
-			-8,
-			-10,
-			-12,
-			-14,
-		},
-
-		Rx1Channel: channelIndexIdentity,
-		Rx1DataRate: func(idx ttnpb.DataRateIndex, offset ttnpb.DataRateOffset, _ bool) (ttnpb.DataRateIndex, error) {
-			if idx > ttnpb.DATA_RATE_7 {
-				return 0, errDataRateIndexTooHigh.WithAttributes("max", 7)
-			}
-			if offset > 5 {
-				return 0, errDataRateOffsetTooHigh.WithAttributes("max", 5)
-			}
-			return downlinkDRTable[idx][offset], nil
-		},
-
-		GenerateChMasks: generateChMask16,
-		ParseChMask:     parseChMask16,
-
-		LoRaCodingRate: "4/5",
-
-		FreqMultiplier:   100,
-		ImplementsCFList: true,
-		CFListType:       ttnpb.CFListType_FREQUENCIES,
-
-		DefaultRx2Parameters: Rx2Parameters{ttnpb.DATA_RATE_0, 869100000},
-
-		Beacon: Beacon{
-			DataRateIndex:    ttnpb.DATA_RATE_3,
-			CodingRate:       "4/5",
-			ComputeFrequency: func(_ float64) uint64 { return 869100000 },
-		},
-		PingSlotFrequency: uint64Ptr(868900000),
-
-		// No LoRaWAN Regional Parameters 1.0
-		// No LoRaWAN Regional Parameters 1.0.1
-		// No LoRaWAN Regional Parameters 1.0.2
-		regionalParameters1_v1_0_3RevA: bandIdentity,
-		regionalParameters1_v1_1RevA:   bandIdentity,
-	}
-	All[RU_864_870] = ru_864_870
-}
+)
